requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
alembic>=1.13
pydantic>=2.0
pydantic-settings>=2.0
PyJWT>=2.8
passlib[bcrypt]>=1.7
bcrypt>=4.0,<4.1
python-multipart>=0.0
openai>=1.0
python-dotenv>=1.0
slowapi>=0.1
pillow>=10.0
boto3>=1.35
